renaming packaging in python library. Resolving module naming conflicts with knack that causes errors
