+++ pylib/modm/azure/arm_template.py
