+++ pylib/modm/azure/client_app.py
