+++ pylib/modm/installer/deployment_type.py
