+++ pylib/modm/installer/package.py
@@ -1,6 +1,6 @@
 import os
 from pathlib import Path
-import packaging._zip_utils as ziputils
+import modm._zip_utils as ziputils
 import shutil
 import tempfile
 from .manifest import ManifestInfo, write_manifest
