+++ pylib/tests/test_azure_arm_template.py
@@ -1,5 +1,5 @@
 import unittest
-from packaging.azure.arm_template import ArmTemplate
+from modm.azure.arm_template import ArmTemplate
 
 
 class TestArmTemplate(unittest.TestCase):
