+++ pylib/tests/test_reserved_template_parameter.py
@@ -1,5 +1,5 @@
 import unittest
-from packaging.installer.reserved_template_parameter import ReservedTemplateParameter, is_reserved
+from modm.installer.reserved_template_parameter import ReservedTemplateParameter, is_reserved
 
 class TestReservedTemplateParameter(unittest.TestCase):
     def test_is_reserved(self):
